refactoring and validations

+++ src/stepstoworks/stepstoworks.service.ts
@@ -1,9 +1,9 @@
-import { HttpStatus, Injectable, ConflictException } from '@nestjs/common';
-import { CreateStepstoworkDto } from './dto/create-stepstowork.dto';
-import { UpdateStepstoworkDto } from './dto/update-stepstowork.dto';
-import { PrismaService } from 'src/prisma/prisma.service';
-import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
-import { ReorderStepstoworkDto } from './dto/reorder-stepstowork.dto';
+import { HttpStatus, Injectable, ConflictException } from "@nestjs/common";
+import { CreateStepstoworkDto } from "./dto/create-stepstowork.dto";
+import { UpdateStepstoworkDto } from "./dto/update-stepstowork.dto";
+import { PrismaService } from "src/prisma/prisma.service";
+import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
+import { ReorderStepstoworkDto } from "./dto/reorder-stepstowork.dto";
 
 @Injectable()
 export class StepstoworksService {
@@ -18,13 +18,13 @@ export class StepstoworksService {
       return {
         statusCode: HttpStatus.CREATED,
         data: item,
-        message: 'The item has been successfully created.',
+        message: "The item has been successfully created.",
       };
     } catch (error) {
       if (error instanceof PrismaClientKnownRequestError) {
-        if (error.code === 'P2002') {
+        if (error.code === "P2002") {
           throw new ConflictException(
-            `Ya existe un item con esta descripción:  ${createStepstoworkDto.description},  en esta orden`,
+            `Ya existe un item con esta descripción:  ${createStepstoworkDto.description},  en esta orden`
           );
         }
       }
@@ -38,7 +38,7 @@ export class StepstoworksService {
         worksId,
       },
       orderBy: {
-        order: 'asc',
+        order: "asc",
       },
       include: {
         user: true,
@@ -50,10 +50,10 @@ export class StepstoworksService {
     return this.prisma.stepsToWork.findMany({
       where: {
         worksId,
-        status: 'PENDING',
+        status: "PENDING",
       },
       orderBy: {
-        order: 'asc',
+        order: "asc",
       },
       include: {
         user: true,
@@ -65,10 +65,10 @@ export class StepstoworksService {
     return this.prisma.stepsToWork.findMany({
       where: {
         worksId,
-        status: 'FINISHED',
+        status: "FINISHED",
       },
       orderBy: {
-        order: 'asc',
+        order: "asc",
       },
       include: {
         user: true,
@@ -80,15 +80,33 @@ export class StepstoworksService {
     // Transacción: actualiza todos los pasos
     const res = await this.prisma.$transaction(
       reorderStepstoworkDto.ordered.map(({ id, order }) =>
-        this.prisma.stepsToWork.update({ where: { id }, data: { order } }),
-      ),
+        this.prisma.stepsToWork.update({ where: { id }, data: { order } })
+      )
     );
     console.log(res);
     return res;
   }
 
-  async editStepToWork(id: string, updateStepstoworkDto: UpdateStepstoworkDto) {
+  async editStepToWork(
+    id: string,
+    workId: string,
+    updateStepstoworkDto: UpdateStepstoworkDto
+  ) {
     try {
+      const statusFinished = await this.prisma.works.findFirst({
+        where: {
+          id: workId,
+          progress: "FINISHED",
+        },
+      });
+      console.log(statusFinished)
+      if (statusFinished) {
+        return {
+          error: "workFinished",
+          message: "La orden está Finalizada, debe cambiar su estatus para editar.",
+        };
+      }
+
       const item = await this.prisma.stepsToWork.update({
         where: { id },
         data: updateStepstoworkDto,
@@ -97,13 +115,13 @@ export class StepstoworksService {
       return {
         statusCode: HttpStatus.CREATED,
         data: item,
-        message: 'The item has been successfully edited.',
+        message: "The item has been successfully edited.",
       };
     } catch (error) {
       if (error instanceof PrismaClientKnownRequestError) {
-        if (error.code === 'P2002') {
+        if (error.code === "P2002") {
           throw new ConflictException(
-            `Ya existe un item con esta descripción:  ${updateStepstoworkDto.description},  en esta orden`,
+            `Ya existe un item con esta descripción:  ${updateStepstoworkDto.description},  en esta orden`
           );
         }
       }
@@ -119,13 +137,39 @@ export class StepstoworksService {
     return `This action updates a #${id} stepstowork`;
   }
 
-  async remove(id: string) {
+  async remove(id: string, idWork: string) {
     try {
+      const check = await this.prisma.stepsToWork.count({
+        where: {
+          worksId: idWork,
+        },
+      });
+
+      const statusFinished = await this.prisma.works.findFirst({
+        where: {
+          id: idWork,
+          progress: "FINISHED",
+        },
+      });
+
+      if (check === 1 && statusFinished) {
+        return {
+          error: "hay un error",
+          message: "No puede eliminar una tarea de una orden Finalizada",
+        };
+      }
+      if (statusFinished) {
+        return {
+          error: "hay un error",
+          message: "No puede eliminar un técnico de una orden finalizada",
+        };
+      }
+
       const item = await this.prisma.stepsToWork.delete({ where: { id } });
       return {
         statusCode: HttpStatus.CREATED,
         data: item,
-        message: 'The item has been successfully deleted.',
+        message: "The item has been successfully deleted.",
       };
     } catch (error) {
       console.log(error);
