+++ src/technicians/technicians.service.ts
@@ -1,8 +1,8 @@
-import { ConflictException, HttpStatus, Injectable } from '@nestjs/common';
-import { CreateTechnicianDto } from './dto/create-technician.dto';
-import { UpdateTechnicianDto } from './dto/update-technician.dto';
-import { PrismaService } from '../prisma/prisma.service';
-import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
+import { ConflictException, HttpStatus, Injectable } from "@nestjs/common";
+import { CreateTechnicianDto } from "./dto/create-technician.dto";
+import { UpdateTechnicianDto } from "./dto/update-technician.dto";
+import { PrismaService } from "../prisma/prisma.service";
+import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
 
 @Injectable()
 export class TechniciansService {
@@ -19,11 +19,11 @@ export class TechniciansService {
       return {
         statusCode: HttpStatus.CREATED,
         data: res,
-        message: 'The technician has been successfully assigned to the work.',
+        message: "The technician has been successfully assigned to the work.",
       };
     } catch (error) {
       if (error instanceof PrismaClientKnownRequestError) {
-        if (error.code === 'P2002') {
+        if (error.code === "P2002") {
           throw new ConflictException(`El técnico ya existe en esta orden`);
         }
       }
@@ -54,13 +54,13 @@ export class TechniciansService {
     try {
       return this.prisma.user.findMany({
         where: {
-          role: 'TECHNICIAN',
+          role: "TECHNICIAN",
           status: {
-            in: ['ACTIVE'],
+            in: ["ACTIVE"],
           },
         },
         orderBy: {
-          updatedAt: 'desc',
+          updatedAt: "desc",
         },
       });
     } catch (error) {
@@ -76,8 +76,42 @@ export class TechniciansService {
     return `This action updates a #${id} technician`;
   }
 
-  remove(idwork: string, idtechnician: string) {
-    return this.prisma.workTechnician.delete({
+  async remove(idwork: string, idtechnician: string) {
+    const check = await this.prisma.workTechnician.count({
+      where:{
+        workId: idwork
+      }
+    });
+    const statusWork = await this.prisma.works.findFirst({
+      where: {
+        id: idwork,
+        progress: "IN_PROGRESS",
+      },
+    });
+
+     const statusFinished = await this.prisma.works.findFirst({
+      where: {
+        id: idwork,
+        progress: "FINISHED",
+      },
+    });
+
+    if (check === 1 && statusWork) {
+      return {
+        error: "hay un error",
+        message:
+          "no puede dejar la orden sin técnico, la debe colocar como 'No Inciada'",
+      };
+    }
+    if (statusFinished) {
+      return {
+        error: "hay un error",
+        message:
+          "No puede eliminar un técnico de una orden finalizada",
+      };
+    }
+
+    return await this.prisma.workTechnician.delete({
       where: {
         workId_technicianId: {
           workId: idwork,
